backend/requirements.txt
requests
fastapi
uvicorn
python-dotenv
pinecone[grpc]
confluent-kafka
openai
fastapi-csrf-protect
redis[hiredis]
slowapi
authlib
bcrypt
pydantic
pyotp
sqlalchemy
boto3
